+++ src/components/CoverArt.tsx
@@ -1,7 +1,7 @@
 import { css, keyframes } from 'goober';
 import { FC, memo } from 'react';
-import { TNowPlayingData } from 'services/now-playing';
 
+import { TNowPlayingData } from 'services/api';
 import { useStoreFocusListeners } from 'services/utils';
 import { screen } from 'services/style';
