+++ src/components/MainIllustration/Face.tsx
@@ -5,7 +5,6 @@ import Layers from './layers';
 
 const Face: FC<{ isHovering: boolean }> = memo(({ isHovering }) => {
   const isInterested = useSiteStore((state) => state.isInterested);
-
   const expression = isInterested
     ? Layers.SURPRISED
     : isHovering
