+++ src/components/MainIllustration/Tree.tsx
@@ -1,103 +1,13 @@
-import { keyframes } from 'goober';
-
 import { useTheme } from 'services/context/theme';
 import { onClickListeners, useHoverListeners } from 'services/utils';
 
+import Layers from './layers';
 import { Group } from './styles';
 
 const LIT_STYLES = {
   fill: '#FFC81A',
 };
 const DIM_STYLES = { fill: '#eee' };
-
-const blink = keyframes`
-  from { fill: #FFC81A; opacity: 0.5; }
-  50% { fill: #FFC81A; opacity: 1; }
-  to { fill: #FFC81A; opacity: 0.5; }
-`;
-
-const BLINKING_LIGHTS = [
-  {
-    d:
-      'M180.713049,99.0088838 C180.713049,112.578696 169.713267,123.578478 156.143455,123.578478 C142.573643,123.578478 131.573861,112.578696 131.573861,99.0088838 C131.573861,85.4390715 142.573643,74.4392895 156.143455,74.4392895 C169.713267,74.4392895 180.713049,85.4390715 180.713049,99.0088838',
-    transform:
-      'translate(156.143455, 99.008884) scale(-1, 1) translate(-156.143455, -99.008884)',
-  },
-  {
-    d:
-      'M399.174813,307.97782 C399.174813,321.547632 388.175031,332.547414 374.605219,332.547414 C361.035407,332.547414 350.035625,321.547632 350.035625,307.97782 C350.035625,294.408008 361.035407,283.408226 374.605219,283.408226 C388.175031,283.408226 399.174813,294.408008 399.174813,307.97782',
-    transform:
-      'translate(374.605219, 307.977820) scale(-1, 1) translate(-374.605219, -307.977820)',
-  },
-  {
-    d:
-      'M107.002953,490.391266 C107.002953,503.961078 96.0031714,514.96086 82.4333591,514.96086 C68.8635469,514.96086 57.8637649,503.961078 57.8637649,490.391266 C57.8637649,476.821453 68.8635469,465.821671 82.4333591,465.821671 C96.0031714,465.821671 107.002953,476.821453 107.002953,490.391266',
-    transform:
-      'translate(82.433359, 490.391266) scale(-1, 1) translate(-82.433359, -490.391266)',
-  },
-  {
-    d:
-      'M448.314659,496.603595 C448.314659,510.173407 437.314877,521.173189 423.745064,521.173189 C410.175252,521.173189 399.17547,510.173407 399.17547,496.603595 C399.17547,483.033782 410.175252,472.034 423.745064,472.034 C437.314877,472.034 448.314659,483.033782 448.314659,496.603595',
-    transform:
-      'translate(423.745064, 496.603595) scale(-1, 1) translate(-423.745064, -496.603595)',
-  },
-  {
-    d:
-      'M150.279269,291.166513 C150.279269,304.736325 139.279487,315.736107 125.709674,315.736107 C112.139862,315.736107 101.14008,304.736325 101.14008,291.166513 C101.14008,277.596701 112.139862,266.596919 125.709674,266.596919 C139.279487,266.596919 150.279269,277.596701 150.279269,291.166513',
-    transform:
-      'translate(125.709674, 291.166513) scale(-1, 1) translate(-125.709674, -291.166513)',
-  },
-  {
-    d:
-      'M293.413275,577.354418 C293.413275,590.92423 282.413493,601.924012 268.843681,601.924012 C255.273869,601.924012 244.274087,590.92423 244.274087,577.354418 C244.274087,563.784606 255.273869,552.784824 268.843681,552.784824 C282.413493,552.784824 293.413275,563.784606 293.413275,577.354418',
-    transform:
-      'translate(268.843681, 577.354418) scale(-1, 1) translate(-268.843681, -577.354418)',
-  },
-  {
-    d:
-      'M49.1391885,714.817286 C49.1391885,728.387098 38.1394065,739.388193 24.5695943,739.388193 C10.999782,739.388193 3.63797881e-12,728.387098 3.63797881e-12,714.817286 C3.63797881e-12,701.248787 10.999782,690.247692 24.5695943,690.247692 C38.1394065,690.247692 49.1391885,701.248787 49.1391885,714.817286',
-    transform:
-      'translate(24.569594, 714.817942) scale(-1, 1) translate(-24.569594, -714.817942)',
-  },
-  {
-    d:
-      'M310.464849,24.5695943 C310.464849,38.1394065 299.465067,49.1391885 285.895254,49.1391885 C272.325442,49.1391885 261.32566,38.1394065 261.32566,24.5695943 C261.32566,10.999782 272.325442,6.60804744e-13 285.895254,6.60804744e-13 C299.465067,6.60804744e-13 310.464849,10.999782 310.464849,24.5695943',
-    transform:
-      'translate(285.895254, 24.569594) scale(-1, 1) translate(-285.895254, -24.569594)',
-  },
-  {
-    d:
-      'M244.273561,479.028633 C244.273561,492.598445 233.273779,503.598227 219.703967,503.598227 C206.134155,503.598227 195.134373,492.598445 195.134373,479.028633 C195.134373,465.45882 206.134155,454.459038 219.703967,454.459038 C233.273779,454.459038 244.273561,465.45882 244.273561,479.028633',
-    transform:
-      'translate(219.703967, 479.028633) scale(-1, 1) translate(-219.703967, -479.028633)',
-  },
-  {
-    d:
-      'M293.413275,243.378527 C293.413275,256.94834 282.413493,267.948122 268.843681,267.948122 C255.273869,267.948122 244.274087,256.94834 244.274087,243.378527 C244.274087,229.808715 255.273869,218.808933 268.843681,218.808933 C282.413493,218.808933 293.413275,229.808715 293.413275,243.378527',
-    transform:
-      'translate(268.843681, 243.378527) scale(-1, 1) translate(-268.843681, -243.378527)',
-  },
-  {
-    d:
-      'M448.314659,723.273828 C448.314659,736.84364 437.314877,747.843422 423.745064,747.843422 C410.175252,747.843422 399.17547,736.84364 399.17547,723.273828 C399.17547,709.704016 410.175252,698.704234 423.745064,698.704234 C437.314877,698.704234 448.314659,709.704016 448.314659,723.273828',
-    transform:
-      'translate(423.745064, 723.273828) scale(-1, 1) translate(-423.745064, -723.273828)',
-  },
-  {
-    d:
-      'M244.273561,754.686768 C244.273561,768.25658 233.273779,779.256362 219.703967,779.256362 C206.134155,779.256362 195.134373,768.25658 195.134373,754.686768 C195.134373,741.116956 206.134155,730.117174 219.703967,730.117174 C233.273779,730.117174 244.273561,741.116956 244.273561,754.686768',
-    transform:
-      'translate(219.703967, 754.686768) scale(-1, 1) translate(-219.703967, -754.686768)',
-  },
-].map((props, i) => (
-  <path
-    {...props}
-    style={{
-      fill: '#FFC81A',
-      animation: `${blink} ${i * 100 + 1000}ms linear infinite`,
-    }}
-  />
-));
 
 const Tree = () => {
   const { isDarkMode, toggleDarkMode } = useTheme();
@@ -172,7 +82,7 @@ const Tree = () => {
           id="Lights"
           transform="translate(493.797468, 672.000000) scale(-1, 1) translate(-493.797468, -672.000000) translate(269.297468, 282.000000)"
         >
-          {BLINKING_LIGHTS}
+          {Layers.BLINKING_LIGHTS}
         </g>
         <g
           id="Star"
