+++ src/components/core/index.ts
@@ -1,4 +1,3 @@
-export * from './Heading';
-export { Page } from './Page';
-export { Text } from './Text';
 export { Link } from './Link';
+export * from './Page';
+export * from './Text';
