+++ src/services/api.ts
@@ -1,6 +1,56 @@
 import { getBestTextColor, ProminentOptions } from 'services/color';
 import { base64Encode } from './utils';
 
+type TGithubStats = {
+  numCommitsSinceLastKnownEvent: number;
+  reposCommittedTo: { name: string; url: string }[];
+};
+
+export const getGithubStats = async (): Promise<TGithubStats | null> => {
+  try {
+    const BASIC_AUTH = base64Encode(`alexieyizhe:${process.env.GITHUB_TOKEN}`);
+
+    const res = await fetch(
+      'https://api.github.com/users/alexieyizhe/events/public?per_page=100&page=1',
+      {
+        headers: {
+          Authorization: `Basic ${BASIC_AUTH}`,
+        },
+      }
+    );
+
+    const data = await res.json();
+    if (!res.ok || !data)
+      throw new Error(`Request error ${res.status}: ${JSON.stringify(data)}`);
+
+    const lastKnownEventDate = new Date(data[data.length - 1].created_at);
+    const pushEvents = data.filter((event: any) => event.type === 'PushEvent');
+    const pushEventRepos = Object.entries<string>(
+      pushEvents.reduce((acc: Record<string, string>, curEvent: any) => {
+        acc[curEvent.repo.name] = `https://github.com/${curEvent.repo.name}`;
+        return acc;
+      }, {})
+    ).map(([name, url]) => ({ name, url }));
+
+    const numCommitsSinceLastKnownEvent = pushEvents.reduce(
+      (acc: number, curEvent: any) => acc + curEvent.payload.size,
+      0
+    );
+
+    if (numCommitsSinceLastKnownEvent === 0) {
+      throw new Error(`No commits since ${lastKnownEventDate.toString()}!`);
+    }
+
+    return {
+      numCommitsSinceLastKnownEvent,
+      reposCommittedTo: pushEventRepos,
+    };
+  } catch (e) {
+    console.error(e);
+    return null;
+  }
+};
+
 export type TNowPlayingData = {
   uri: string;
   type: 'episode' | 'track';
@@ -17,7 +67,7 @@ export const isNowPlayingData = (
 ): status is TNowPlayingData =>
   typeof status === 'object' && !!status.name && !!status.coverArtSrc;
 
-export const requestNewToken = async () => {
+export const requestNewSpotifyToken = async () => {
   const ENCODED_CLIENT_INFO = base64Encode(
     `${process.env.CLIENT_ID}:${process.env.CLIENT_SECRET}`
   );
